Gets us within margin of error perf with styled-components and emotion!

```
┌──────────────────────┬───────────────────┬───────────────────┬─────────────────────────┬───────────────────┬──────────────────────┐
│ Benchmark            │          Avg time │     vs use-styles │ vs use-styles-published │        vs emotion │ vs styled-components │
├──────────────────────┼───────────────────┼───────────────────┼─────────────────────────┼───────────────────┼──────────────────────┤
│ use-styles           │ 22.18ms - 22.50ms │                   │                  faster │            slower │               unsure │
│                      │                   │          -        │                9% - 12% │           0% - 2% │            -0% - +2% │
│                      │                   │                   │         2.34ms - 2.96ms │   0.04ms - 0.50ms │    -0.08ms - +0.40ms │
├──────────────────────┼───────────────────┼───────────────────┼─────────────────────────┼───────────────────┼──────────────────────┤
│ use-styles-published │ 24.73ms - 25.26ms │            slower │                         │            slower │               slower │
│                      │                   │         10% - 13% │                -        │         12% - 15% │            11% - 14% │
│                      │                   │   2.34ms - 2.96ms │                         │   2.61ms - 3.23ms │      2.49ms - 3.13ms │
├──────────────────────┼───────────────────┼───────────────────┼─────────────────────────┼───────────────────┼──────────────────────┤
│ emotion              │ 21.91ms - 22.24ms │            faster │                  faster │                   │               unsure │
│                      │                   │           0% - 2% │               11% - 13% │          -        │            -2% - +1% │
│                      │                   │   0.04ms - 0.50ms │         2.61ms - 3.23ms │                   │    -0.35ms - +0.13ms │
├──────────────────────┼───────────────────┼───────────────────┼─────────────────────────┼───────────────────┼──────────────────────┤
│ styled-components    │ 22.01ms - 22.36ms │            unsure │                  faster │            unsure │                      │
│                      │                   │         -2% - +0% │               10% - 12% │         -1% - +2% │             -        │
│                      │                   │ -0.40ms - +0.08ms │         2.49ms - 3.13ms │ -0.13ms - +0.35ms │                      │
└──────────────────────┴───────────────────┴───────────────────┴─────────────────────────┴───────────────────┴──────────────────────┘
```